+++ app/src/main/AndroidManifest.xml
@@ -26,6 +26,7 @@
                 <category android:name="android.intent.category.DEFAULT" />
             </intent-filter>
         </activity>
+        <service android:name=".MyService"  android:label="@string/app_name" android:enabled="true"/>
     </application>
 
 </manifest>
